refactor: Explodes Scribe
