+++ packages/prophet/Scribe/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { default } from "./Scribe";
