+++ packages/prophet/index.js
@@ -18,9 +18,9 @@ export { default as Prophet } from "./api/Prophet";
 export type Transaction = Object;
 
 export { default as Oracle } from "./Oracle";
+export { default as Scribe } from "./Scribe";
+
 export { default as AuthorityNexus } from "./prophet/AuthorityNexus";
 export { default as DecoderArray } from "./prophet/DecoderArray";
 export { default as FalseProphet } from "./prophet/FalseProphet";
 export { default as FalseProphetDiscourse } from "./prophet/FalseProphetDiscourse";
-export { default as Scribe } from "./prophet/Scribe";
-export { default as ScribePartitionConnection } from "./prophet/ScribePartitionConnection";
